Avoid having all commands and folders in each binary file, remove old hexen.c and hexendd.c
