+++ build.bat
@@ -1,4 +1,4 @@
 gcc blood.c -O3 -o dosbox_bloodgdx.exe
-gcc heretic.c -O3 -DHERETIC
-gcc hexen.c -O3 -o dosbox_chocolate-hexen.exe
-gcc hexendd.c -O3 -o dosbox_chocolate-hexendd.exe
+gcc choco.c -O3 -DHERETIC -o dosbox_chocolate-heretic.exe
+gcc choco.c -O3 -DHEXEN -o dosbox_chocolate-hexen.exe
+gcc choco.c -O3 -DHEXENDD -o dosbox_chocolate-hexendd.exe
