--- hexen.c
@@ -1,5 +1,0 @@
-#include <stdlib.h>
-
-int main(void) {
-    system(".\\chocolate-hexen\\chocolate-hexen.exe -iwad .\\base\\HEXEN.WAD -fullscreen");
-}
